### 2021-06-15 #3373 @whedon
[PRE REVIEW]: Datum: Creating and loading TFRecord datasets
**Submitting author:** @n3011 (<a href="http://orcid.org/0000-0002-0059-2822">Mrinal Haloi</a>)
**Repository:** <a href="https://github.com/openAGI/datum" target ="_blank">https://github.com/openAGI/datum</a>
**Version:** 1.1.0
