### 2021-06-30 #3373 @pmeier
@jmschrei In general, yes. I don't have any experience with `tensorflow` or the `tfrecords` data / file format though. Any review that I could provide would be limited to how the API of `datum` wraps these two. Maybe @n3011 can give some context if that would be enough.
