### 2021-06-30 #3373 @jmschrei
Personally, I think that it would be valuable to get feedback from someone outside the immediate use-case (i.e., tensorflow/tfrecords) as long as it's from someone in the general area (deep learning in Python). But yes, @n3011 do you think that @pmeier would have sufficient expertise to review this?
