### 2021-06-30 #3373 @pmeier
> do you think that @pmeier would have sufficient expertise to review this?

To give a little more context, I am the maintainer of [PyTorch's vision datasets](https://github.com/pytorch/vision). Thus, if no specific knowledge about `tensorflow` / `tfrecords` is required, I'm confident I can provide good feedback.
