### 2021-06-30 #3373 @n3011
@jmschrei  This project is developed specifically for TensorFlow developers. A reviewer with experience in Tensorflow would have been nice.  It solves some teething issues related to creating and loading tfrecords files.
@pmeier Just to give some context, Tensorflow uses the tfrecord format to store dataset samples in binary protocol buffer format, which reduces read overheads when dealing with thousands of files for large datasets.  But creating and loading tfrecord files involve writing a lot of code. `Datum` is developed to minimize the efforts required to create and load TFRecord dataset. Also, it is tightly integrated with `tf.data` API, which makes it easier to deal with large datasets for distributed training.

Overall, I also agree that a deep learning developer with a slight understanding of Tensorflow can also review this.
