### 2021-07-02 #3446 @jmschrei
@n3011 is there a paper for this submission? Is it possible to put it in a "paper" folder in the root directory?
