### 2021-07-09 #3446 @n3011
> @n3011 is there a paper for this submission? Is it possible to put it in a "paper" folder in the root directory?

This issue is resolved, right?
