### 2021-07-09 #3446 @pmeier
> This issue is resolved, right?

The paper is available on the [`joss` branch](https://github.com/openAGI/datum/tree/joss).

That leaves the question which version / branch should be reviewed:
- `master`
- `joss`
- release `1.1.0`
?
