### 2021-07-09 #3446 @n3011
`master` branch has the code of the recent release `1.2.0`. Please review `master` branch.
