### 2021-07-15 #3446 @pmeier
This is my preliminary review based on commit [3878ec9](https://github.com/openAGI/datum/tree/3878ec9ac042e761626f9b3d3b9ecc40dd848317).

## Installation

Installation with `pip` it successful. I couldn't find any mention of supported Python versions. I've assumed 3.7, because this version is used in the development environment.

## Development

I couldn't find any documentation on how to run the linting and tests. I've assumed the commands extracted from the CI workflows. Both linting and tests are passing. For tests script reports that it will run the tests in parallel but doesn't. The test runner caught one warning pointing to `datum` that could be fixed.

## Documentation

I find the documentation insufficient. There are only two documents: the README and one other file in the `docs/` folder. The latter is not linked in the README and thus will only be found by looking through the source. Both contain some examples, but miss an API documentation. Without diving into the source, I didn't understand the interplay between the different classes.

Due to this, I was not able to try `datum` beyond the given examples. I'll hold my review of the functionality for now.

## Statement of Need / State of the field

The README and the paper claim that `datum` is able to reduce the amount of boilerplate code to write and read `TFRecords` datasets. Without significant effort it is impossible to validate that, since a side-by-side comparison is missing.

Furthermore, I couldn't find any mention of [`tensorflow-datasets`](https://www.tensorflow.org/datasets/). Just like `tensorflow`, this library is developed by Google and also provides functionalities to create `TFRecord` datasets. Thus, a comparison should probably use this as baseline rather than vanilla `tensorflow`.
