### 2021-07-26 openAGI/datum#8 @younglululu
Installation failure in either pip install or conda install
**Short description**
Description of the bug.

**Environment information**
* Operating System: ubuntu 18.04
* All python/tensorflow version follow environment.yml

** Error info using conda install

conda env create -f environment.yml
Collecting package metadata: done
Solving environment: done
==> WARNING: A newer version of conda exists. <==
  current version: 4.6.14
  latest version: 4.10.3
Please update conda by running
    $ conda update -n base conda
Preparing transaction: done
Verifying transaction: done
Executing transaction: done
Ran pip subprocess with arguments:
[u'/home/yanglu/anaconda3/envs/datum/bin/python', '-m', 'pip', 'install', '-U', '-r', '/home/yanglu/Documents/datum/tmpXTxcEC.requirements.txt']
Pip subprocess output:
Collecting tensorflow==2.5.0 (from -r file:requirements.txt (line 2))
  Using cached https://files.pythonhosted.org/packages/aa/fd/993aa1333eb54d9f000863fe8ec61e41d12eb833dea51484c76c038718b5/tensorflow-2.5.0-cp37-cp37m-manylinux2010_x86_64.whl
Collecting opencv-python==4.5.2.52 (from -r file:requirements.txt (line 3))
Pip subprocess error:
  ERROR: Could not find a version that satisfies the requirement opencv-python==4.5.2.52 (from -r file:requirements.txt (line 3)) (from versions: 3.4.2.17, 3.4.3.18, 3.4.4.19, 3.4.5.20, 3.4.6.27, 3.4.7.28, 3.4.8.29, 3.4.9.31, 3.4.9.33, 3.4.10.37, 3.4.11.39, 3.4.11.41, 3.4.11.43, 3.4.11.45, 3.4.13.47, 3.4.15.55, 4.0.0.21, 4.0.1.23, 4.0.1.24, 4.1.0.25, 4.1.1.26, 4.1.2.30, 4.2.0.32, 4.2.0.34, 4.3.0.38, 4.4.0.40, 4.4.0.42, 4.4.0.44, 4.4.0.46, 4.5.1.48, 4.5.3.56)
ERROR: No matching distribution found for opencv-python==4.5.2.52 (from -r file:requirements.txt (line 3))
WARNING: You are using pip version 19.1.1, however version 21.2.1 is available.
You should consider upgrading via the 'pip install --upgrade pip' command.
CondaEnvException: Pip failed

** Error info using pip install

Collecting datum
  Using cached datum-1.2.0.tar.gz (40 kB)
     command: /home/yanglu/anaconda3/bin/python -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-pFVc1m/datum/setup.py'"'"'; __file__='"'"'/tmp/pip-install-pFVc1m/datum/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /tmp/pip-pip-egg-info-SEXAtA
         cwd: /tmp/pip-install-pFVc1m/datum/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-install-pFVc1m/datum/setup.py", line 9, in <module>
        from setuptools import find_namespace_packages, setup
    ImportError: cannot import name find_namespace_packages
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
