### 2021-07-26 openAGI/datum#8 @n3011
Which python version are you using?
This package works with python >= 3.7

You can install by running `pip install datum`
