### 2021-07-26 openAGI/datum#8 @younglululu
Yes, I am using python 3.7 by creating a new conda environ suggested by environment.yml.

And pip install failed as well. :(
