### 2021-07-27 openAGI/datum#8 @n3011
You don't need to build an environment using the `environment.yml` file when you install datum using `pip`.

If you are getting errors, that means that your development environment is not set up correctly.
