### 2021-07-27 openAGI/datum#8 @jmschrei
My understanding is that using conda to set up the environment, as you did, should always create a working environment even if your base one is incompatible with the prereqs. That's the primary motivation behind having virtual environments at all, right? What happens when you create a new blank environment Py3.7 environment and try pip installing into that?
