### 2021-07-31 #3446 @pmeier
@jmschrei I did my preliminary review in https://github.com/openjournals/joss-reviews/issues/3446#issuecomment-880581078. Since then there have been no commits, so I'm on hold for now.
