+++ datum/export/export.py
@@ -31,31 +31,34 @@ def export_to_tfrecord(input_path: str, output_path: str, problem_type: str,
     input_path: Root path to input data folder.
     output_path: Path to store output tfrecords and generated metadata.
     problem_type: Type of the problem, see `datum.probelm.types` for available problems.
+    write_configs: Configuration for tfrecord writing.
 
     write_configs has the following configurable attributes:
-      generator: Generator class.
-      serializer: Serializer instance.
-      splits: A dict with split names as keys and split attributes as values.
-        Following split attributes are supported:
-          num_examples: Number of examples in the split.
-          extension: Input image extension in case of image data, all input should have same
-            extension.
-            default: For image data, `.jpg`
-          image_dir: Name of the directory containing the data, used for image classification.
-            default: split name, eg: train for classification, for detection `JPEGImages` -a s per
-              VOC12 folder structure.
-          csv_path: Path to ground truths csv file, used for classification dataset.
-            default: split name with .csv extension, eg: train.csv
-          set_dir: In case of VOC12 style  datasets, image set information.
-            default: `ImageSets` as per VOC12 dataset folder structure
-          annotation_dir: Directory with annotations, used for VOC12 style datasets.
-            default: `Annotations` as per VOC12 dataset folder structure
-          label_dir: Directory with label images, used in segmentation.
-            default: `SegmentationClass` for as per VOC12 folder structure.
-          image_extension: Extension of input images, used in segmentation.
-            default: `.jpg`
-          label_extension: Extension of label images, used in segmentation.
-            default: `.png`
+
+    generator: Generator class.
+    serializer: Serializer instance.
+    splits: A dict with split names as keys and split attributes as values.
+
+    Following split attributes are supported:
+
+    num_examples: Number of examples in the split.
+    extension: Input image extension in case of image data, all input should have same
+      extension. Default - For image data, `.jpg`
+    image_dir: Name of the directory containing the data, used for image classification.
+      Default - split name, `train` for classification, for detection `JPEGImages` as per
+        VOC12 folder structure.
+    csv_path: Path to ground truths csv file, used for classification dataset.
+      Default - split name with .csv extension, example - `train.csv`
+    set_dir: In case of VOC12 style  datasets, image set information.
+      Default - `ImageSets` as per VOC12 dataset folder structure
+    annotation_dir: Directory with annotations, used for VOC12 style datasets.
+      Default - `Annotations` as per VOC12 dataset folder structure
+    label_dir: Directory with label images, used in segmentation.
+      Default - `SegmentationClass` for as per VOC12 folder structure.
+    image_extension: Extension of input images, used in segmentation.
+      Default - `.jpg`
+    label_extension: Extension of label images, used in segmentation.
+      Default - `.png`
 
     Raises:
       ValueError: If splits information is not in the dict format.
