+++ mkdocs.yml
@@ -1,0 +1,2 @@
+site_name: Datum
+theme: readthedocs

[…]
