+++ docs/api-docs/encoder.md
@@ -1,0 +1,16 @@
+<!-- markdownlint-disable -->
+
+<a href="../../datum/encoder/__init__.py#L0"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+# <kbd>module</kbd> `encoder`
+
+
+
+
+
+
+
+
+---
+
+_This file was automatically generated via [lazydocs](https://github.com/ml-tooling/lazydocs)._

[…]
