+++ setup.py
@@ -8,22 +8,22 @@ import sys
 
 from setuptools import find_namespace_packages, setup
 
-project_name = 'datum'
+project_name = "datum"
 
-version_path = os.path.join(os.path.dirname(__file__), 'datum')
+version_path = os.path.join(os.path.dirname(__file__), "datum")
 sys.path.append(version_path)
 from version import __version__ # isort:skip
 
-with open('README.md') as fp:
+with open("README.md") as fp:
   _LONG_DESCRIPTION = fp.read()
 
 REQUIRED_PKGS = []
 
-with open('requirements.txt', 'r') as f:
+with open("requirements.txt", "r") as f:
   ignore = 0
   for tentative_req in f.readlines():
     if tentative_req:
-      if tentative_req.startswith('#'):
+      if tentative_req.startswith("#"):
         ignore += 1
         if ignore > 1:
           break
@@ -31,8 +31,8 @@ with open('requirements.txt', 'r') as f:
         REQUIRED_PKGS.append(tentative_req.strip())
 
 TESTS_REQUIRE = [
-    'pytest',
-    'pytest-xdist',
+    "pytest",
+    "pytest-xdist",
 ]
 packages = find_namespace_packages(exclude=["tests*", "tools*"],)
 
@@ -43,20 +43,21 @@ setup(
     "Datum provides APIs to create tfrecord daatsets and read tfrecord as tf.data.Datasets",
     long_description=
     "Datum provides APIs to create tfrecord daatsets and read tfrecord as tf.data.Datasets",
-    author='OpenAGI',
-    author_email='maintainer@openagi.io',
-    url='https://github.com/openagi/datum',
-    download_url='https://github.com/openagi/datum/tags',
-    license='Apache 2.0',
+    author="OpenAGI",
+    author_email="maintainer@openagi.io",
+    url="https://github.com/openagi/datum",
+    download_url="https://github.com/openagi/datum/tags",
+    license="Apache 2.0",
     packages=packages,
     scripts=[],
     include_package_data=True,
     install_requires=REQUIRED_PKGS,
+    python_requires=">=3.7",
     tests_require=TESTS_REQUIRE,
     classifiers=[
-        'Intended Audience :: Developers',
-        'Intended Audience :: Science/Research',
-        'License :: OSI Approved :: Apache Software License',
+        "Intended Audience :: Developers",
+        "Intended Audience :: Science/Research",
+        "License :: OSI Approved :: Apache Software License",
     ],
-    keywords='datum tensorflow tf.data datasets tfrecord',
+    keywords="datum tensorflow tf.data datasets tfrecord",
 )
