### 2021-08-01 openAGI/datum#8 @pmeier
@younglululu Are you on macOS? If so, which version?

By looking at the files associated with the PyPI releases, the following architectures are supported by the three most recent releases of `opencv-python`:

- [`4.5.1.48`](https://pypi.org/project/opencv-python/4.5.1.48/#files): `macosx_10_13_x86_64`
- [`4.5.2.52`](https://pypi.org/project/opencv-python/4.5.2.52/#files): `macosx_10_15_x86_64`
- [`4.5.3.56`](https://pypi.org/project/opencv-python/4.5.3.56/#files): `macosx_10_15_x86_64`, `macosx_11_0_arm64` (only Python 3.7 and higher)

@n3011 If there is no specific reason for the `opencv==4.5.2.52` pin, I suggest you remove or relax it.
