### 2021-08-01 openAGI/datum#8 @pmeier
@younglululu Are you an ARM machine?

By looking at the files associated with the PyPI releases, `opencv-python==4.5.2.52` was not released for ARM:

- [`4.5.1.48`](https://pypi.org/project/opencv-python/4.5.1.48/#files): `manylinux2014_aarch64`, `manylinux2014_i686`, `manylinux2014_x86_64`
- [`4.5.2.52`](https://pypi.org/project/opencv-python/4.5.2.52/#files): `manylinux2014_x86_64`
- [`4.5.3.56`](https://pypi.org/project/opencv-python/4.5.3.56/#files): `manylinux2014_aarch64`, `manylinux2014_x86_64`

@n3011 If there is no specific reason for the pin, I suggest you remove or relax it.
