### 2021-08-01 openAGI/datum#8 @n3011
@pmeier I will relax the requirements to `4.5.*`
