+++ requirements.txt
@@ -1,6 +1,6 @@
 # core lib dep
 tensorflow==2.5.0
-opencv-python==4.5.2.52
+opencv-python==4.5.*
 absl-py>=0.10
 tqdm>=4.39.0
 
