### 2021-08-01 openAGI/datum#8 @younglululu
My environment is Ubuntu 18.04, not mac.
I don't think I have an ARM machine, it should be conventional windows-intel architecture.
