+++ paper.bib
@@ -15,6 +15,13 @@
   	Year = 2016
 }
 
+@book{tfds,
+  	url = {https://www.tensorflow.org/datasets},
+  	Author = {Google},
+    Title = {{TensorFlow Datasets}, A collection of ready-to-use datasets},
+  	Year = 2018
+}
+
 @article{datum,
     author = {{Mrinal Haloi}},
     title = "{Datum}",
