### 2021-08-02 #3446 @n3011
@jmschrei  I have updated the project and the `paper.md` file to address comments from @pmeier .

Package related changes:
1. Issue related to running tests in parallel is fixed and now in the master branch.
2. API-doc strings are added.

In the `joss` branch, `paper.md` and `paper.bib` file is updated to address the followings:
1. Information about TensorFlow-datasets. Tensorflow datasets is designed for publicly available datasets. To use it with custom datasets users still need to dataset-specific interface.
2. Included a para on the complexities involved in converting dataset to tfrecord format
