### 2021-08-02 #3446 @pmeier
@n3011

> To use it with custom datasets users still need to dataset-specific interface.

They have [extensive documentation](https://www.tensorflow.org/datasets/add_dataset) about how to add a custom dataset. From my current perspective they provide a superset of what `datum` provides.

To convince me that this is not the case and `datum` indeed reduces the amount of boilerplate code needed to store and load a `TFRecord` dataset, I would like to see side-by-side comparison. That means, given the dataset structure you chose for `datum` what does the code look like to do it with `datum` and `tfds`.
