### 2021-08-03 #3446 @jmschrei
@n3011 thanks for responding to the feedback. I think that @pmeier's suggestion that you compare against TFDS is reasonable, and I would also like to see it. Please include such a comparison in your next revisions. Thanks!
