### 2021-08-04 #3446 @pmeier
@n3011

> I will consider that as out of scope for this paper.

Given that `datum`'s central claim is "Storing and saving `TFRecord` datasets is hard and tedious. `datum` helps with that.", providing evidence for this claim is very much in scope.

IMO, a side-by-side comparison is the easiest way to do that, but you are not bound to that. If you find another way to do such a comparison, I'm happy to review it.
