+++ requirements.txt
@@ -1,5 +1,5 @@
 # core lib dep
-tensorflow==2.5.0
+tensorflow==2.5.1
 opencv-python==4.5.*
 absl-py>=0.10
 tqdm>=4.39.0
