### 2021-09-20 #3446 @younglululu
This is my view of datum:

I followed the instruction and failed at the installation stage by using either pip install or conda install.

I created an issue regarding the failure (https://github.com/openAGI/datum/issues/8), and it was closed without solving the problem.

There is no way I can review this paper without the successful installation.
