### 2021-09-20 #3446 @pmeier
@younglululu openAGI/datum#8 should have been resolved by https://github.com/openAGI/datum/commit/b7f34ceedad48d03e50feb2b0e7a14ae353af0b5, which is included in the latest `datum==1.2.1` release. Have you retried the installation since then? If there is still some error, would you mind creating a new issue with the log output?
